requirements.txt
streamlit-1.44.0-py3-none-any.whl